ensuring naming convention tests review singular test 2011 switching numeric to float casting due to truncating errors adding credit cards dimension

+++ dbt_adventure_works/models/staging/erp/stg_erp__product.sql
@@ -13,11 +13,11 @@ with
         -- Product Information          
         ,cast(name as varchar) as product_name
         ,finishedgoodsflag as finished_goods_flag
-        ,cast(standardcost as numeric) as standard_cost
-        ,cast(listprice as numeric) as unit_price
+        ,cast(standardcost as float) as standard_cost
+        ,cast(listprice as float) as unit_price
         ,cast(productline as varchar) as product_line
         ,cast(productmodelid as string) as product_model_id
-        , 'size' as product_size
+        ,'size' as product_size
         ,cast(color as varchar) as product_color
         ,cast(sellstartdate as date) as sell_start_date
         ,cast(sellenddate as date) as sell_end_date
